% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{build_url}
\alias{build_url}
\title{Build a Medicare dataset resource request URL}
\usage{
build_url(dataset_id)
}
\arguments{
\item{dataset_id}{Socrata dataset identifier}
}
\value{
Full API URL for dataset request
}
\description{
Build a Medicare dataset resource request URL
}
