% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the user's Socrata API token}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{character string represnting the user's Socrata app token.}
}
\description{
Set the user's Socrata API token
}
\examples{
\dontrun{set_api_key("yourkey")}
}
