% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{socrata_api_key}
\alias{socrata_api_key}
\title{Set the user's Socrata API token}
\usage{
socrata_api_key(api_key)
}
\arguments{
\item{api_key}{character string represnting the user's Socrata API token.}
}
\description{
Set the user's Socrata API token
}
