% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve Medicare dataset}
\usage{
get_data(dataset_id, app_token = NULL)
}
\arguments{
\item{dataset_id}{Socrata dataset identifier}

\item{app_token}{Optional Socrata application token}
}
\value{
R data frame with dataset
}
\description{
Retrieve Medicare dataset
}
